restrict unauthorized Edit Delete

+++ base/views.py
@@ -1,5 +1,7 @@
 from django.shortcuts import render, redirect
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
 from django.db.models import Q  # Q allows to use and/or to filter objects
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
@@ -54,6 +56,7 @@ def room (request, pk):
 
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(request):
     if request.method == 'POST':
         form = RoomForm(request.POST)
@@ -65,8 +68,13 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id = pk)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here!!")
+    
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -77,8 +85,13 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id = pk)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here!!")
+    
     if request.method == 'POST':
         room.delete()
         return redirect('home')
